requirements.txt
ffmpeg-python>=0.2.0
psutil>=5.9.0
pyinstaller>=5.13.0
PySide6>=6.7.0
requests>=2.31.0

